修改totalCount类型,添加Pageable接口并改造

+++ myframe-dao/src/test/java/com/myframe/dao/TestUserDao.java
@@ -5,5 +5,5 @@ import com.myframe.pojo.TestUser;
 
 public interface TestUserDao extends MyBatisGenericDao<TestUser> {
 
-    Long getCount2();
+    Integer getCount2();
 }
